% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeType.R
\name{changeType}
\alias{changeType}
\title{Change variable type
\code{changeType} change variable type across list of data frame}
\usage{
changeType(x, Var, funct = "as.Date")
}
\arguments{
\item{x}{list of data frames}

\item{Var}{variables or column names}

\item{funct}{function to change variable type such as.factor(), as.Date(), as.factor(), etc}
}
\value{
list of data frames with changed variable type
}
\description{
Change variable type
\code{changeType} change variable type across list of data frame
}
\examples{
# make a list
iris_list <- list(iris1 = iris, iris2 = iris)

# change one variable type
iris_list <- lapply(iris_list, changeType, Var = "Sepal.Width", funct = "as.character")

# change 2 variables type
iris_list <- lapply(iris_list, changeType, Var = c("Sepal.Length", "Species"), funct = "as.character")
}
