% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histNA_byVar.R
\name{histNA_byVar}
\alias{histNA_byVar}
\title{Histogram for NAs}
\usage{
histNA_byVar(dat, NAvar, byvar, bin = 30)
}
\arguments{
\item{dat}{data frame}

\item{NAvar}{variable with NAs}

\item{byvar}{influencing variable}

\item{bin}{number of bins (default is 30)}
}
\value{
two histogram of influencing variable which separated by the presence of NAs in the NAvar
}
\description{
\code{histNA_byVar} check the distribution of NAs of certain variable is affected by another variable
}
\examples{
dat <- iris
dat[dat$Species == "setosa", "Sepal.Length"] <- NA
#indicate right-tailed missingness
histNA_byVar(dat, Sepal.Length, Sepal.Width)
}
