% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histWithCurve.R
\name{histWithCurve}
\alias{histWithCurve}
\title{Histogram with curve
Histogram with a normal density curve}
\usage{
histWithCurve(x, breaks = 24)
}
\arguments{
\item{x}{the vector or column name of a dataset}

\item{breaks}{number of bins}
}
\value{

}
\description{
Histogram with curve
Histogram with a normal density curve
}
\examples{
histWithCurve(iris$Sepal.Length)
}
