% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read all excel sheets
Read all excel sheets}
\usage{
read_excel_allsheets(filename, pages = sheets, tibble = FALSE)
}
\arguments{
\item{filename}{excel files}

\item{pages}{number of order of sheets in the excel (to select certain sheets)}

\item{tibble}{set to TRUE for tibble format (default is FALSE)}
}
\value{

}
\description{
Read all excel sheets
Read all excel sheets
}
\examples{
# Read all excel sheets
# read_excel_allsheets("datasets")

# Read several excel sheets
# read_excel_allsheets("datasets", pages = 2:5)
}
