% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDiag.R
\name{regDiag}
\alias{regDiag}
\title{Regression diagnostic for outlier and influential case
These function is only for screening}
\usage{
regDiag(model)
}
\arguments{
\item{model}{lm or glm model}
}
\value{
Outlier : leverage, Studentized deleted residual (SDR)

Influential case : DFFits, DFBetas, cook's distance (cook.d)

also noted that we use dfbetas (scaled/standardized) NOT dfbeta function

TRUE : outlier and influential case present,

FALSE : no outlier and influential case

Cut-off values:

leverage > 0.2

SDR >= 3 or <= -3

DFFits > 1

DFBetas > 1

cook.d > 0.5
}
\description{
Regression diagnostic for outlier and influential case
These function is only for screening
}
\examples{
mod <- lm(Sepal.Length ~ Species + Sepal.Width, data = iris)
regDiag(mod)
}
